Add: Add functionality to handle POST request for updating long URL
This commit adds a POST request handler in express_server.js to update the long URL in the urlDatabase for the specified ID. It uses the route '/urls/:id' to receive the request and updates the corresponding entry in urlDatabase. After updating, the user is redirected back to the '/urls' page.

- Added POST route '/urls/:id' handler in express_server.js
- Updated the long URL in urlDatabase for the specified ID
- Redirected the user to '/urls' page after updating.