update: Implement URL Filtering and Error Handling in /urls Routes
Updated the /urls GET route to display only URLs created by the logged-in user using the urlsForUser function.
Implemented error handling to display a relevant message if no URLs are found for the current user.
Added conditionals to the /urls/:id GET route to display an error message if no user is logged in and if the user tries to edit a shortURL that belongs to another user.